+++ src/chart/animator/options.h
@@ -16,6 +16,8 @@ namespace Anim
 class Enum(SectionId)
 	(style,title,legend,show,hide,color,coordSystem,geometry,y,x,tooltip);
 
+class Enum(RegroupStrategy)(fade,drilldown,aggregate);
+
 class Options
 {
 public:
@@ -25,12 +27,16 @@ public:
 		std::optional<::Anim::Duration> delay;
 		std::optional<::Anim::Duration> duration;
 		void set(const std::string &param, const std::string &value);
+		bool isSet() const;
 	};
 
 	struct Keyframe {
+		std::optional<RegroupStrategy> regroupStrategy;
 		Section all;
 		std::array<Section, SectionId::EnumInfo::count()> sections;
+		Section &get(SectionId sectionId);
 		const Section &get(SectionId sectionId) const;
+		RegroupStrategy getRegroupStrategy() const;
 	};
 
 	struct Control {