+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 
 ### Added
 
+- 'regroupStrategy' animation option introduced to control the algorithm for 
+  transitioning between charts having the data grouped differently on them 
+  (containing a different set of categorical dataseries).
 - Multi keyframe animation support (one animation through multiple chart).
 - 'style' property returns the style object only filled with the user-set 
   values, all the values (returned by this property till 0.6.x) can be get